Update react/context example - Now more in line with conventional react context example
